fix(project): separate a project card component to individual components
