+++ src/entities/project/ui/_shared/index.ts
@@ -1,1 +1,2 @@
 export { ProjectModalBase } from './ProjectModalBase';
+export { ProjectCard, type ProjectCardData } from './ProjectCard';
