+++ src/entities/project/ui/index.ts
@@ -1,2 +1,1 @@
 export * from './ProjectEmptyState';
-export * from './ProjectCard';
