+++ src/entities/project/ui/project-card/index.ts
@@ -1,0 +1,3 @@
+export * from './ProjectCardButton';
+export * from './ProjectCardDates';
+export * from './ClockIcon';
